rename abbreviations.lua to autocorrect
