+++ lua/config/init.lua
@@ -3,4 +3,4 @@ require("config.lazy")
 require("config.colors").set("rose-pine")
 require("config.options")
 require("config.tmux_vim_nav")
-require("config.abbreviations")
+require("config.autocorrect")
